renderPoint вынесена в utils/point.js
