+++ src/main.js
@@ -1,14 +1,13 @@
-import {render, RenderPosition, replace} from './utils/render.js';
+import {render, RenderPosition} from './utils/render.js';
 import PointsListView from './view/points-list.js';
 import TripInfoView from './view/trip-info.js';
 import Menu from './view/menu.js';
 import FiltersForm from './view/filters-form.js';
 import SortForm from './view/sort-form.js';
 import {getPointsList} from './mock/point.js';
-import PointView from './view/point.js';
-import EditFormView from './view/edit-event-form.js';
 import NoPoints from './view/no-points.js';
-import {generateID, isEsc} from './utils/common';
+import {generateID} from './utils/common.js';
+import {renderPoint} from './utils/point.js';
 
 const POINTS_COUNT = 20;
 
@@ -39,39 +38,5 @@ render(tripEventsElement, pointsListComponent, RenderPosition.BEFOREEND);
 render(tripNavigationElement, menuComponent, RenderPosition.BEFOREEND);
 
 const pointsListElement = tripEventsElement.querySelector('.trip-events__list');
-
-const renderPoint = (container, point) => {
-  const pointComponent = new PointView(point);
-  const editFormComponent = new EditFormView(point);
-
-  const replacePointToForm = () => replace(editFormComponent, pointComponent);
-  const replaceFormToPoint = () => replace(pointComponent, editFormComponent);
-
-  const onDocumentEscKeydown = (evt) => {
-    if (isEsc(evt)) {
-      evt.preventDefault();
-      replaceFormToPoint();
-      document.removeEventListener('keydown', onDocumentEscKeydown);
-    }
-  };
-
-  pointComponent.setRollupBtnClickHandler(() => {
-    replacePointToForm();
-    document.addEventListener('keydown', onDocumentEscKeydown);
-  });
-
-  editFormComponent.setFormSubmitHandler((evt) => {
-    evt.preventDefault();
-    replaceFormToPoint();
-    document.removeEventListener('keydown', onDocumentEscKeydown);
-  });
-
-  editFormComponent.setResetBtnClickHandler(() => {
-    replaceFormToPoint();
-    document.removeEventListener('keydown', onDocumentEscKeydown);
-  });
-
-  render(container, pointComponent, RenderPosition.BEFOREEND);
-};
 
 points.forEach((point) => renderPoint(pointsListElement, point));
